+++ src/auth/auth.module.ts
@@ -1,0 +1,31 @@
+import { MiddlewareConsumer, Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { LoggerMiddleware } from '@app/middleware/logger.middleware';
+import { JwtModule } from '@nestjs/jwt';
+import { ConfigService } from '@nestjs/config';
+import { UsersModule } from '@app/users/users.module';
+
+@Module({
+  imports: [
+    UsersModule,
+    JwtModule.registerAsync({
+      useFactory: (config: ConfigService) => {
+        return {
+          secret: config.get<string>('auth.secret_key'),
+          signOptions: {
+            expiresIn: config.get<string | number>('auth.expiration_time'),
+          },
+        };
+      },
+      inject: [ConfigService],
+    }),
+  ],
+  controllers: [AuthController],
+  providers: [AuthService],
+})
+export class AuthModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(LoggerMiddleware).forRoutes(AuthController);
+  }
+}