+++ src/credentials/dto/create-credential.dto.ts
@@ -1,0 +1,15 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class CreateCredentialDto {
+  @ApiProperty({
+    description: 'The email of the user',
+    example: 'john@doe.com',
+  })
+  email: string;
+
+  @ApiProperty({
+    description: 'The password of the user',
+    example: 'password123',
+  })
+  password: string;
+}