+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ValidationPipe } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from '../prisma/prisma.service';
@@ -8,23 +8,56 @@ import { User } from '@prisma/client';
 export class UsersService {
   constructor(private prisma: PrismaService) {}
 
-  create(createUserDto: CreateUserDto) {
-    return this.prisma.user.create({ data: createUserDto });
+  async create(createUserDto: CreateUserDto): Promise<User> {
+    const validationPipe = new ValidationPipe({ transform: true });
+    await validationPipe.transform(createUserDto, { type: 'body' });
+    const user = await this.prisma.user.create({
+      data: {
+        firstName: createUserDto.firstName,
+        lastName: createUserDto.lastName,
+        citizenId: createUserDto.idCard,
+        email: createUserDto.email,
+        phoneNumber: createUserDto.phone,
+        password: createUserDto.password,
+        displayName: createUserDto.nickname,
+        gender: createUserDto.gender,
+        dateOfBirth: createUserDto.dateOfBirth,
+        address: createUserDto.address,
+        city: createUserDto.city,
+        zipcode: createUserDto.zipcode,
+      },
+    });
+
+    return user;
   }
 
   async findAll(): Promise<User[]> {
-    return this.prisma.user.findMany();
+    return await this.prisma.user.findMany();
   }
 
-  findOne(userId: string) {
-    return this.prisma.user.findUnique({ where: { userId } });
+  async findOne(userId: string): Promise<User> {
+    return await this.prisma.user.findUniqueOrThrow({
+      where: { userId: userId },
+    });
   }
 
-  update(userId: string, updateUserDto: UpdateUserDto) {
-    return this.prisma.user.update({ where: { userId }, data: updateUserDto });
+  //looks up database for users with the given email
+  //and returns true if there are no such users
+  async findUserWithEmail(email: string): Promise<User | null> {
+    const user = await this.prisma.user.findFirst({
+      where: { email: email },
+    });
+    return user;
   }
 
-  remove(userId: string) {
-    return this.prisma.user.delete({ where: { userId } });
+  async update(userId: string, updateUserDto: UpdateUserDto): Promise<User> {
+    return await this.prisma.user.update({
+      where: { userId: userId },
+      data: updateUserDto,
+    });
+  }
+
+  async remove(userId: string): Promise<User> {
+    return await this.prisma.user.delete({ where: { userId } });
   }
 }